Before queue_policy
